Fix up header, add router, switch component locations
