+++ index.html
@@ -9,7 +9,7 @@
 		<meta property="og:type" content="" />
 		<meta property="og:url" content="" />
 		<meta property="og:image" content="" />
-		<title>FF7 No-Slots for St. Jude #4</title>
+		<title>FF7 No-Slots for St. Jude</title>
 		<link rel="stylesheet" href="./assets/dist/css/styles.css" />
 	</head>
